+++ cookbooks/kafka-bcpc/metadata.rb
@@ -1,10 +1,10 @@
 # encoding: utf-8
 
 name             'kafka-bcpc'
-maintainer       'Unsupportd'
-description      'Ad-Hoc Recipe to setup prerequisites for Kafka server'
+maintainer       'Bloomberg Finance L.P.'
+description      'Recipes to setup prerequisites for Kafka cluster'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.0.0'
+version          '0.1.0'
 
 depends "kafka"
 depends "sysctl"